Grease the skids for ktor based testing
This commit introduces Ktor to provide a testing server for cache updates. It also reduces the expect/actual footprint
by minimizing the code that is platform dependent.

modified:   build.gradle.kts
- Add ktor dependencies.
modified:   src/commonMain/kotlin/com/pajato/tmdb/lib/DatasetManager.kt
- Add a baseUrl parameter to support local server based testing.
modified:   src/commonMain/kotlin/com/pajato/tmdb/lib/TmdbData.kt
new file:   src/commonMain/kotlin/com/pajato/tmdb/lib/TmdbFetch.kt
- Move support artifacts to TmdbFetch.kt
modified:   src/commonTest/kotlin/com/pajato/tmdb/lib/LibraryTest.kt
- Minor edits.
deleted:    src/jvmMain/kotlin/com/pajato/tmdb/lib/TmdbDataJVM.kt
new file:   src/jvmMain/kotlin/com/pajato/tmdb/lib/TmdbFetchJVM.kt
modified:   src/jvmTest/kotlin/com/pajato/tmdb/lib/LibraryTestJVM.kt
- Start coroutine testing and adding ktor support.
